+++ components/browser/state/src/main/java/mozilla/components/browser/state/reducer/TabListReducer.kt
@@ -68,15 +68,11 @@ internal object TabListReducer {
                 if (tabTarget == null) {
                     state
                 } else {
-                    val targetPosition = state.tabs.indexOf(tabTarget) + (if (action.placeAfter) 1 else 0)
-                    val positionOffset = state.tabs.filterIndexed { index, tab ->
-                        (index < targetPosition && tab.id in action.tabIds)
-                    }.count()
-                    val finalPos = targetPosition - positionOffset
-                    val (movedTabs, unmovedTabs) = state.tabs.partition { it.id in action.tabIds }
-                    val updatedTabList = unmovedTabs.subList(0, finalPos) +
-                        movedTabs +
-                        unmovedTabs.subList(finalPos, unmovedTabs.size)
+                    val targetPosition = state.tabs.indexOf(tabTarget)
+                    val (movedTab, unmovedTabs) = state.tabs.partition { it.id == action.tabId }
+                    val updatedTabList = unmovedTabs.subList(0, targetPosition) +
+                        movedTab +
+                        unmovedTabs.subList(targetPosition, unmovedTabs.size)
 
                     state.copy(tabs = updatedTabList)
                 }