+++ components/feature/tabs/src/main/java/mozilla/components/feature/tabs/TabsUseCases.kt
@@ -485,23 +485,20 @@ class TabsUseCases(
         private val store: BrowserStore
     ) {
         /**
-         * Moves the tabs of [tabIds] next to [targetTabId], before/after based on [placeAfter]
+         * Moves a tab into a new position
          *
-         * @property tabIds The IDs of the tabs to move.
-         * @property targetTabId A tab that the moved tabs will be moved next to
-         * @property placeAfter True if the moved tabs should be placed after the target,
-         * False for placing before the target. Irrelevant if the target is one of the tabs being moved,
-         * since then the whole list is moved to where the target was. Ordering of the moved tabs
-         * relative to each other is preserved.
+         * @property tabId The IDs of the tab to move.
+         * @property targetTabId The moved tab will replace the position of the target.
+         * Before/After determined by the original position of the moved tab,
+         * pushing the target towards the original tab's position.
          */
         operator fun invoke(
-            tabIds: List<String>,
+            tabId: String,
             targetTabId: String,
-            placeAfter: Boolean
         ) {
             store.dispatch(
                 TabListAction.MoveTabsAction(
-                    tabIds, targetTabId, placeAfter
+                    tabId, targetTabId
                 )
             )
         }